Update to the last Rust: remove float.

+++ src/broad/brute_force_bounding_volume_broad_phase.rs
@@ -223,14 +223,14 @@ mod test {
 
     #[test]
     fn test_bfbv_empty() {
-        type Shape = WithAABB<Vec3<float>, Ball<float>>;
+        type Shape = WithAABB<Vec3<f64>, Ball<f64>>;
         let dispatcher: NoIdDispatcher<Shape> = NoIdDispatcher;
         let mut bf     = BruteForceBoundingVolumeBroadPhase::new(dispatcher, 0.2);
         let ball       = Ball::new(0.3);
 
         for i in range(-10, 10) {
             for j in range(-10, 10) {
-                bf.add(@mut WithAABB(Vec3::new(i as float * 30.0, j as float * 30.0, 0.0), ball));
+                bf.add(@mut WithAABB(Vec3::new(i as f64 * 30.0, j as f64 * 30.0, 0.0), ball));
             }
         }
 
@@ -241,7 +241,7 @@ mod test {
 
     #[test]
     fn test_bfbv_nbh_collide() {
-        type Shape = WithAABB<Vec3<float>, Ball<float>>;
+        type Shape = WithAABB<Vec3<f64>, Ball<f64>>;
         let dispatcher: NoIdDispatcher<Shape> = NoIdDispatcher;
         let mut bf     = BruteForceBoundingVolumeBroadPhase::new(dispatcher, 0.2);
         let ball       = Ball::new(0.3);
@@ -249,7 +249,7 @@ mod test {
         // create a grid
         for i in range(-10, 10) {
             for j in range(-10, 10) {
-                bf.add(@mut WithAABB(Vec3::new(i as float * 0.9, j as float * 0.9, 0.0), ball));
+                bf.add(@mut WithAABB(Vec3::new(i as f64 * 0.9, j as f64 * 0.9, 0.0), ball));
             }
         }
 
@@ -266,7 +266,7 @@ mod test {
 
     #[test]
     fn test_dbvt_nbh_move_collide() {
-        type Shape = WithAABB<Vec3<float>, Ball<float>>;
+        type Shape = WithAABB<Vec3<f64>, Ball<f64>>;
         let dispatcher: NoIdDispatcher<Shape> = NoIdDispatcher;
         let mut bf     = BruteForceBoundingVolumeBroadPhase::new(dispatcher, 0.2);
         let ball       = Ball::new(0.3);
@@ -276,7 +276,7 @@ mod test {
         // create a grid
         for i in range(-10, 10) {
             for j in range(-10, 10) {
-                let to_add = @mut WithAABB(Vec3::new(i as float * 0.9, j as float * 0.9, 0.0), ball);
+                let to_add = @mut WithAABB(Vec3::new(i as f64 * 0.9, j as f64 * 0.9, 0.0), ball);
                 bf.add(to_add);
                 to_move.push(to_add);
             }
@@ -300,7 +300,7 @@ mod test {
 
     #[test]
     fn test_bfbv_quadratic_collide() {
-        type Shape = WithAABB<Vec3<float>, Ball<float>>;
+        type Shape = WithAABB<Vec3<f64>, Ball<f64>>;
         let dispatcher: NoIdDispatcher<Shape> = NoIdDispatcher;
         let mut bf     = BruteForceBoundingVolumeBroadPhase::new(dispatcher, 0.2);
         let ball       = Ball::new(0.3);
